+++ src/main/java/org/stellar/sdk/federation/exception/MalformedAddressException.java
@@ -1,6 +1,8 @@
 package org.stellar.sdk.federation.exception;
 
-import org.stellar.sdk.exception.SdkException;
-
 /** Given Stellar address is malformed. */
-public class MalformedAddressException extends SdkException {}
+public class MalformedAddressException extends IllegalArgumentException {
+  public MalformedAddressException(String message) {
+    super(message);
+  }
+}